requirements.txt
Django
psycopg2-binary
requests
pillow
python-decouple
django-cors-headers
dj-database-url
djangorestframework
djangorestframework-simplejwt
django-stubs
drf-spectacular
pytest
pytest-django
schemathesis
